Refactoring Map.js
