+++ packages/web-frontend/src/containers/Map/Map.js
@@ -6,34 +6,15 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import styled from 'styled-components';
-import { createSelector } from 'reselect';
-import {
-  TileLayer,
-  MarkerLayer,
-  LeafletMap,
-  InteractivePolygon,
-  ZoomControl,
-} from '@tupaia/ui-components/lib/map';
-
-import { checkBoundsDifference, organisationUnitIsArea } from '../../utils';
+import { TileLayer, LeafletMap } from '@tupaia/ui-components/lib/map';
+import { checkBoundsDifference } from '../../utils';
 import { DemoLand } from './DemoLand';
 import { DisasterLayer } from './DisasterLayer';
-import {
-  selectActiveTileSet,
-  selectMeasuresWithDisplayInfo,
-  selectCurrentMapOverlayCodes,
-  selectCurrentOrgUnit,
-  selectHasPolygonMeasure,
-  selectMeasureOptions,
-  selectOrgUnit,
-  selectOrgUnitChildren,
-  selectOrgUnitSiblings,
-  selectRenderedMeasuresWithDisplayInfo,
-  selectAreRegionLabelsPermanent,
-  selectMeasureData,
-} from '../../selectors';
-import { changePosition, closeDropdownOverlays, setOrgUnit, setMobileTab } from '../../actions';
+import { ZoomControl } from './ZoomControl';
+import { selectActiveTileSet, selectCurrentMapOverlayCodes } from '../../selectors';
+import { changePosition, closeDropdownOverlays, setOrgUnit } from '../../actions';
 import { TRANS_BLACK, TRANS_BLACK_LESS } from '../../styles';
+import { DataVisualsLayer } from './DataVisualsLayer/DataVisualsLayer';
 
 const CHANGE_TO_PARENT_PERCENTAGE = 0.6;
 
@@ -74,40 +55,15 @@ class MapComponent extends Component {
   }
 
   shouldComponentUpdate(nextProps) {
-    const {
-      currentOrganisationUnit,
-      displayedChildren,
-      mapOverlayCodes,
-      serieses,
-      measureData,
-      position,
-      tileSetUrl,
-    } = this.props;
+    const { mapOverlayCodes, position, sidePanelWidth, tileSetUrl } = this.props;
 
     if (JSON.stringify(nextProps.mapOverlayCodes) !== JSON.stringify(mapOverlayCodes)) {
-      return true;
-    }
-
-    // Only updates/re-renders when the measure has changed or the orgUnit has changed.
-    // These are the only cases where polygons or area tooltips should rerender.
-    if (JSON.stringify(nextProps.displayedChildren) !== JSON.stringify(displayedChildren)) {
-      return true;
-    }
-
-    if (
-      nextProps.currentOrganisationUnit?.organisationUnitCode !==
-      currentOrganisationUnit?.organisationUnitCode
-    ) {
       return true;
     }
 
     if (nextProps.tileSetUrl !== tileSetUrl) return true;
 
     if (JSON.stringify(nextProps.position) !== JSON.stringify(position)) return true;
-
-    if (JSON.stringify(nextProps.serieses) !== JSON.stringify(serieses)) return true;
-
-    if (JSON.stringify(nextProps.measureData) !== JSON.stringify(measureData)) return true;
 
     return false;
   }
@@ -145,40 +101,7 @@ class MapComponent extends Component {
   }
 
   render() {
-    const {
-      onCloseDropdownOverlays,
-      currentOrganisationUnitSiblings,
-      currentOrganisationUnit,
-      displayedChildren,
-      getChildren,
-      measureData,
-      onChangeOrgUnit,
-      onSeeOrgUnitDashboard,
-      serieses,
-      multiOverlayMeasureData,
-      multiOverlaySerieses,
-      position,
-      shouldSnapToPosition,
-      tileSetUrl,
-      measureOrgUnits,
-      permanentLabels,
-      showAttribution,
-      showZoomControl,
-    } = this.props;
-
-    // Only show data with valid coordinates. Note: this also removes region data
-    const processedData = measureData.filter(
-      ({ coordinates }) => coordinates && coordinates.length === 2,
-    );
-    const hasMeasureData = measureData && measureData.length > 0;
-    const basicPropsForInteractivePolygon = {
-      hasMeasureData,
-      measureOrgUnits,
-      multiOverlaySerieses,
-      multiOverlayMeasureData,
-      onChangeOrgUnit,
-      permanentLabels,
-    };
+    const { onCloseDropdownOverlays, position, shouldSnapToPosition, tileSetUrl } = this.props;
 
     return (
       <StyledMap
@@ -193,38 +116,7 @@ class MapComponent extends Component {
         <TileLayer tileSetUrl={tileSetUrl} showAttribution={showAttribution} />
         {showZoomControl && <ZoomControl position="bottomright" />}
         <DemoLand />
-        {currentOrganisationUnit && organisationUnitIsArea(currentOrganisationUnit) && (
-          <InteractivePolygon
-            area={currentOrganisationUnit}
-            organisationUnitChildren={getChildren(currentOrganisationUnit.organisationUnitCode)}
-            isActive
-            {...basicPropsForInteractivePolygon}
-          />
-        )}
-        {displayedChildren?.map(area => (
-          <InteractivePolygon
-            area={area}
-            key={area.organisationUnitCode}
-            organisationUnitChildren={getChildren(area.organisationUnitCode)}
-            isChildArea
-            {...basicPropsForInteractivePolygon}
-          />
-        ))}
-        {currentOrganisationUnitSiblings?.map(area => (
-          <InteractivePolygon
-            area={area}
-            key={area.organisationUnitCode}
-            organisationUnitChildren={getChildren(area.organisationUnitCode)}
-            {...basicPropsForInteractivePolygon}
-          />
-        ))}
-        <MarkerLayer
-          measureData={processedData}
-          serieses={serieses || null}
-          onSeeOrgUnitDashboard={onSeeOrgUnitDashboard}
-          multiOverlayMeasureData={multiOverlayMeasureData}
-          multiOverlaySerieses={multiOverlaySerieses}
-        />
+        <DataVisualsLayer />
         <DisasterLayer />
       </StyledMap>
     );
@@ -235,16 +127,7 @@ MapComponent.propTypes = {
   onCloseDropdownOverlays: PropTypes.func.isRequired,
   onChangePosition: PropTypes.func.isRequired,
   currentParent: PropTypes.object,
-  currentOrganisationUnit: PropTypes.object.isRequired,
-  currentOrganisationUnitSiblings: PropTypes.array.isRequired,
-  displayedChildren: PropTypes.arrayOf(PropTypes.object),
-  getChildren: PropTypes.func.isRequired,
-  measureData: PropTypes.array,
-  measureOrgUnits: PropTypes.array,
   mapOverlayCodes: PropTypes.array.isRequired,
-  serieses: PropTypes.array,
-  multiOverlayMeasureData: PropTypes.array,
-  multiOverlaySerieses: PropTypes.array,
   position: PropTypes.shape({
     center: PropTypes.oneOfType([PropTypes.object, PropTypes.array]),
     bounds: PropTypes.oneOfType([PropTypes.object, PropTypes.array]),
@@ -254,86 +137,24 @@ MapComponent.propTypes = {
   onSeeOrgUnitDashboard: PropTypes.func.isRequired,
   shouldSnapToPosition: PropTypes.bool.isRequired,
   tileSetUrl: PropTypes.string.isRequired,
-  permanentLabels: PropTypes.bool,
-  showZoomControl: PropTypes.bool,
-  showAttribution: PropTypes.bool,
 };
 
 MapComponent.defaultProps = {
-  displayedChildren: [],
-  measureData: [],
-  measureOrgUnits: [],
-  serieses: [],
-  multiOverlayMeasureData: [],
-  multiOverlaySerieses: [],
   currentParent: null,
-  permanentLabels: undefined,
-  showZoomControl: true,
-  showAttribution: true,
 };
 
-const selectMeasureDataWithCoordinates = createSelector([measureData => measureData], measureData =>
-  measureData.map(({ location, ...otherData }) => ({
-    ...otherData,
-    coordinates: location && location.point,
-    region: location && location.region,
-  })),
-);
-
 const mapStateToProps = state => {
-  const { isAnimating, shouldSnapToPosition, position, displayedMapOverlays } = state.map;
+  const { shouldSnapToPosition, position } = state.map;
   const mapOverlayCodes = selectCurrentMapOverlayCodes(state);
-  const currentOrganisationUnit = selectCurrentOrgUnit(state);
-  const currentParent = selectOrgUnit(state, currentOrganisationUnit.parent);
-  const currentChildren =
-    selectOrgUnitChildren(state, currentOrganisationUnit.organisationUnitCode) || [];
-  const measureData = selectMeasureDataWithCoordinates(
-    selectRenderedMeasuresWithDisplayInfo(state, displayedMapOverlays),
-  );
-  const serieses = selectMeasureOptions(state, displayedMapOverlays);
-  const permanentLabels = selectAreRegionLabelsPermanent(state);
-  const multiOverlayMeasureData = selectMeasureData(state, mapOverlayCodes);
-  const multiOverlaySerieses = selectMeasureOptions(state, mapOverlayCodes);
-  // If the org unit's grandchildren are polygons and have a measure, display grandchildren
-  // rather than children
-  let displayedChildren = currentChildren;
-  let measureOrgUnits = [];
-
-  if (selectHasPolygonMeasure(state)) {
-    measureOrgUnits = selectMeasuresWithDisplayInfo(state, displayedMapOverlays);
-    const measureOrgUnitCodes = measureOrgUnits.map(orgUnit => orgUnit.organisationUnitCode);
-    const grandchildren = currentChildren
-      .map(area => selectOrgUnitChildren(state, area.organisationUnitCode))
-      .reduce((acc, val) => acc.concat(val), []); // equivelent to .flat(), for IE
-
-    const hasShadedGrandchildren =
-      grandchildren &&
-      grandchildren.some(child => measureOrgUnitCodes.includes(child.organisationUnitCode));
-    if (hasShadedGrandchildren) displayedChildren = grandchildren;
-  }
-
-  const getChildren = organisationUnitCode => selectOrgUnitChildren(state, organisationUnitCode);
+  const { isSidePanelExpanded } = state.global;
+  const { contractedWidth, expandedWidth } = state.dashboard;
 
   return {
-    position,
-    currentOrganisationUnit,
-    currentParent,
-    displayedChildren,
-    measureData,
-    serieses,
     mapOverlayCodes,
-    measureOrgUnits,
-    multiOverlayMeasureData,
-    multiOverlaySerieses,
-    currentOrganisationUnitSiblings: selectOrgUnitSiblings(
-      state,
-      currentOrganisationUnit.organisationUnitCode,
-    ),
-    getChildren,
+    position,
     tileSetUrl: selectActiveTileSet(state).url,
-    isAnimating,
     shouldSnapToPosition,
-    permanentLabels,
+    sidePanelWidth: isSidePanelExpanded ? expandedWidth : contractedWidth,
   };
 };
